+++ src/solo12_driver.cpp
@@ -10,6 +10,8 @@
 #include <yaml-cpp/yaml.h>
 #include <boost/range/adaptor/indexed.hpp>
 
+#include <robot_interfaces_solo/solo12_utils.hpp>
+
 namespace robot_interfaces_solo
 {
 Solo12Driver::Solo12Driver(const Solo12Config &config) : config_(config)
@@ -171,59 +173,6 @@ void Solo12Driver::shutdown()
     }
 }
 
-template <typename T>
-void _set_optional_config_value(const YAML::Node &user_config,
-                                const std::string &name,
-                                T *var)
-{
-    try
-    {
-        if (user_config[name])
-        {
-            *var = user_config[name].as<T>();
-        }
-    }
-    catch (const YAML::Exception &e)
-    {
-        throw std::runtime_error(fmt::format(
-            "FATAL: Failed to load parameter '{}' from configuration file: {}",
-            name,
-            e.what()));
-    };
-}
-
-Solo12Config Solo12Config::from_file(
-    const std::filesystem::path &config_file_name)
-{
-    Solo12Config config;
-    YAML::Node user_config;
-
-    try
-    {
-        user_config = YAML::LoadFile(config_file_name);
-    }
-    catch (const YAML::Exception &e)
-    {
-        throw std::runtime_error(
-            fmt::format("FATAL: Failed to load configuration from '{}': {}",
-                        config_file_name.string(),
-                        e.what()));
-    }
-
-    _set_optional_config_value(
-        user_config, "network_interface", &config.network_interface);
-    _set_optional_config_value(
-        user_config, "slider_serial_port", &config.slider_serial_port);
-    _set_optional_config_value(
-        user_config, "max_motor_current_A", &config.max_motor_current_A);
-    _set_optional_config_value(
-        user_config, "home_offset_rad", &config.home_offset_rad);
-    _set_optional_config_value(
-        user_config, "logger_level", &config.logger_level);
-
-    return config;
-}
-
 FakeSolo12Driver::FakeSolo12Driver(const Solo12Config &config) : config_(config)
 {
     // initialise logger and set level based on config
@@ -307,32 +256,18 @@ void FakeSolo12Driver::shutdown()
 {
 }
 
-Solo12Backend::Ptr create_solo12_backend(Solo12Data::Ptr robot_data,
-                                         const Solo12Config &driver_config,
-                                         const double first_action_timeout,
-                                         const uint32_t max_number_of_actions)
+Solo12Backend::Ptr create_real_solo12_backend(
+    Solo12Data::Ptr robot_data,
+    const Solo12Config &driver_config,
+    const double first_action_timeout,
+    const uint32_t max_number_of_actions)
 {
-    constexpr double MAX_ACTION_DURATION_S = 0.003;
-    constexpr double MAX_INTER_ACTION_DURATION_S = 0.005;
-
-    // config.print();
-
-    // wrap the actual robot driver directly in a MonitoredRobotDriver
-    auto monitored_driver =
-        std::make_shared<robot_interfaces::MonitoredRobotDriver<Solo12Driver>>(
-            std::make_shared<Solo12Driver>(driver_config),
-            MAX_ACTION_DURATION_S,
-            MAX_INTER_ACTION_DURATION_S);
-
-    constexpr bool real_time_mode = true;
-    auto backend = std::make_shared<Solo12Backend>(monitored_driver,
-                                                   robot_data,
-                                                   real_time_mode,
-                                                   first_action_timeout,
-                                                   max_number_of_actions);
-    backend->set_max_action_repetitions(std::numeric_limits<uint32_t>::max());
-
-    return backend;
+    constexpr bool enable_timing_watchdog = true;
+    return create_solo12_backend(robot_data,
+                                 std::make_shared<Solo12Driver>(driver_config),
+                                 first_action_timeout,
+                                 max_number_of_actions,
+                                 enable_timing_watchdog);
 }
 
 Solo12Backend::Ptr create_fake_solo12_backend(
@@ -343,15 +278,12 @@ Solo12Backend::Ptr create_fake_solo12_backend(
 {
     auto driver = std::make_shared<FakeSolo12Driver>(driver_config);
 
-    constexpr bool real_time_mode = true;
-    auto backend = std::make_shared<Solo12Backend>(driver,
-                                                   robot_data,
-                                                   real_time_mode,
-                                                   first_action_timeout,
-                                                   max_number_of_actions);
-    backend->set_max_action_repetitions(std::numeric_limits<uint32_t>::max());
-
-    return backend;
+    constexpr bool enable_timing_watchdog = false;
+    return create_solo12_backend(robot_data,
+                                 driver,
+                                 first_action_timeout,
+                                 max_number_of_actions,
+                                 enable_timing_watchdog);
 }
 
 }  // namespace robot_interfaces_solo