+++ srcpy/solo12.cpp
@@ -13,6 +13,8 @@
 
 #include <robot_interfaces/pybind_helper.hpp>
 #include <robot_interfaces_solo/solo12_driver.hpp>
+#include <robot_interfaces_solo/solo12_pybullet_driver.hpp>
+#include <robot_interfaces_solo/solo12_utils.hpp>
 
 namespace ris = robot_interfaces_solo;
 
@@ -132,8 +134,30 @@ PYBIND11_MODULE(solo12, m)
              "Load configuration from a YAML file (using default values for "
              "parameters missing in the file).");
 
+    pybind11::class_<ris::BaseSolo12Driver, ris::BaseSolo12Driver::Ptr>
+        base_driver(m, "BaseSolo12Driver");
+
+    pybind11::class_<ris::PyBulletSolo12Driver,
+                     std::shared_ptr<ris::PyBulletSolo12Driver>,
+                     ris::BaseSolo12Driver>(m, "PyBulletDriver")
+        .def(pybind11::init<bool, bool, bool, const std::string &>(),
+             pybind11::arg("real_time_mode") = true,
+             pybind11::arg("visualize") = true,
+             pybind11::arg("use_fixed_base") = false,
+             pybind11::arg("logger_level") = "debug")
+        .def("get_bullet_env", &ris::PyBulletSolo12Driver::get_bullet_env);
+
     m.def("create_backend",
           &ris::create_solo12_backend,
+          pybind11::arg("robot_data"),
+          pybind11::arg("robot_driver"),
+          pybind11::arg("first_action_timeout") =
+              std::numeric_limits<double>::infinity(),
+          pybind11::arg("max_number_of_actions") = 0,
+          pybind11::arg("enable_timing_watchdog") = true);
+
+    m.def("create_real_backend",
+          &ris::create_real_solo12_backend,
           pybind11::arg("robot_data"),
           pybind11::arg("driver_config"),
           pybind11::arg("first_action_timeout") =
@@ -142,6 +166,14 @@ PYBIND11_MODULE(solo12, m)
 
     m.def("create_fake_backend",
           &ris::create_fake_solo12_backend,
+          pybind11::arg("robot_data"),
+          pybind11::arg("driver_config"),
+          pybind11::arg("first_action_timeout") =
+              std::numeric_limits<double>::infinity(),
+          pybind11::arg("max_number_of_actions") = 0);
+
+    m.def("create_pybullet_backend",
+          &ris::create_pybullet_solo12_backend,
           pybind11::arg("robot_data"),
           pybind11::arg("driver_config"),
           pybind11::arg("first_action_timeout") =